+++ docs/_pages/releases.md
@@ -11,6 +11,7 @@ sidebar:
 
 ### What's new
 
+* Added extensive support for `Span<T>` / `ReadOnlySpan<T>` and `Be(string)` for `Span<char>` / `ReadOnlySpan<char>` - [#3172](https://github.com/fluentassertions/fluentassertions/pull/3172)
 * Added `Excluding<T>()` and `Excluding(Type)` overloads to exclude all members of a certain type from equivalency comparisons - [#3115](https://github.com/fluentassertions/fluentassertions/pull/3115)
 * `BeEquivalentTo` will now report the missing or extraneous items for differently sized collections - [#3133](https://github.com/fluentassertions/fluentassertions/pull/3133)
 * Added `HaveMillisecond` and `NotHaveMillisecond` assertion methods for `DateTime` and `DateTimeOffset` - [#3164](https://github.com/fluentassertions/fluentassertions/pull/3164)